Rename Match classes for PHP 8 compatibility
