+++ lib/Doctrine/ODM/MongoDB/Aggregation/Builder.php
@@ -394,9 +394,9 @@ class Builder
      *
      * @see https://docs.mongodb.com/manual/reference/operator/aggregation/match/
      */
-    public function match() : Stage\Match
+    public function match() : Stage\MatchStage
     {
-        $stage = new Stage\Match($this);
+        $stage = new Stage\MatchStage($this);
         $this->addStage($stage);
 
         return $stage;
